refactored project folder to 'amrvac_pytools'
