+++ amrvac_pytools/vtkfiles/numpy_support.py
@@ -29,7 +29,7 @@ Created by Prabhu Ramachandran in Feb. 2008.
 import vtk
 import numpy
 
-from amrvac_tools.vtkfiles import vtkConstants
+from amrvac_pytools.vtkfiles import vtkConstants
 
 # Useful constants for VTK arrays.
 VTK_ID_TYPE_SIZE = vtk.vtkIdTypeArray().GetDataTypeSize()
